benchmark/requirements.txt
awscli==1.16.106
boto3==1.9.96
botocore==1.12.96
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
decorator==4.3.0
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.8.2
docutils==0.14
idna==2.7
ijson==2.3
ipython-genutils==0.2.0
iso8601==0.1.12
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.4.0
nbformat==4.4.0
pyasn1==0.4.5
pycryptodome==3.6.6
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
retrying==1.3.3
rsa==3.4.2
s3transfer==0.2.0
six==1.11.0
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0